+++ assets/src/scripts/blocks/song-difficulty/BlockSave.jsx
@@ -1,0 +1,12 @@
+import { __ } from '@wordpress/i18n';
+
+export const BlockSave = ({ className, attributes }) => {
+    const { difficulty } = attributes;
+
+    return (
+        <div className={className}>
+            {difficulty && <span>{__('Song difficulty: ')}</span>}
+            {difficulty && <span>{difficulty}</span>}
+        </div>
+    )
+}