+++ assets/src/scripts/blocks/youtube-post-template/BlockSave.jsx
@@ -1,12 +1,21 @@
 const { InnerBlocks } = wp.blockEditor;
 import { SidebarCta } from '../../components/SidebarCta';
 
-export const BlockSave = ({ attributes, className }) => {
+/**
+ * Save component for Youtube Post Template.
+ * 
+ * @param {object} props Gutenberg default props.
+ * @returns {JSX}
+ */
+export const BlockSave = ({ attributes }) => {
     const {
+        songTitle,
         videoUrlEmbed,
         videoInfoFetched,
         sidebarCtaSlotOne,
-        sidebarCtaSlotTwo
+        sidebarCtaSlotTwo,
+        songChordsAttribute,
+        songDifficultyAttribute,
     } = attributes;
 
     return (
@@ -25,6 +34,14 @@ export const BlockSave = ({ attributes, className }) => {
                     { sidebarCtaSlotTwo.id > 0 && <SidebarCta cta={sidebarCtaSlotTwo} />}
                 </div> }
             </div>
+            <div className="whatup">{ songDifficultyAttribute }</div>
+            {songChordsAttribute && <div className="song-chords">
+                <span className="song-chords-title">{`${songTitle} Chords:`}</span>
+                <ul className="song-chords-list">{songChordsAttribute.map(chord => <li className="chord">{chord}</li>)}</ul>
+            </div>}
+            {songDifficultyAttribute && <div className="song-difficulty">
+                <span className="song-difficulty-title">{`${songTitle} Difficulty: ${songDifficultyAttribute}`}</span>
+            </div>}
             <div class={`youtube-post-type-video-description`}>
                 <InnerBlocks.Content />
             </div>