+++ assets/src/scripts/blocks/ypt-search/BlockEdit.jsx
@@ -1,5 +1,13 @@
 const { __ } = wp.i18n;
+const { serverSideRender: ServerSideRender } = wp;
 
+/**
+ * Placeholder block edit that shows what the component will look like on the frontend.
+ * 
+ * Block's Save component is rendered in PHP in views/blocks/ypt-search.php
+ * 
+ * @returns {JSX}
+ */
 export const BlockEdit = () => (
-    <h2>{__('A song/youtube post search box will appear here.')}</h2>
+    <ServerSideRender block="gutenberg-good-guitarist/ypt-search" />
 )