+++ inc/Core/Tags.php
@@ -46,10 +46,7 @@ class Tags
 		// Hide category and tag text for pages.
 		if ('post' === get_post_type()) {
 			/* translators: used between list items, there is a space after the comma */
-			$categories_list = get_the_category_list(esc_html__(', ', 'awps'));
-			if ($categories_list && self::categorized_blog()) {
-				printf('<span class="cat-links">'.esc_html__('Posted in %1$s', 'awps').'</span>', $categories_list); // WPCS: XSS OK.
-			}
+			$categories_list = get_the_category_list(esc_html__(', '));
 			/* translators: used between list items, there is a space after the comma */
 			$tags_list = get_the_tag_list('', esc_html__(', ', 'awps'));
 			if ($tags_list) {