+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### Sept 2, 2022
+- Center post titles
+- Add song chords and difficulty blocks for YPT block
+- Add new Feel taxonomy
+- Add additional code comments
+- Remove "Posted in Uncategorized" from appearing in post footer
+
 ### July 27, 2022
 - Fix song title and artist not appearing on latest lessons carousel